pyproject.toml
[project]
name = "FootySim-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "pydantic-settings>=2.3",
  "aiosqlite>=0.20",
  "footysim-core>=0.1.7",  #  ma lib
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}

[tool.uvicorn]
app-dir = "src"
